+++ processor/src/test/fixtures/expected/com/example/cascade_dispose/Arez_DisposeNotifierDisabledCascadeDisposeModel.java
@@ -1,4 +1,4 @@
-package com.example.pre_dispose;
+package com.example.cascade_dispose;
 
 import arez.Arez;
 import arez.ArezContext;
@@ -10,17 +10,17 @@ import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 
 @Generated("arez.processor.ArezProcessor")
-final class Arez_MultiPreDisposeNotDisposeNotifierModel extends MultiPreDisposeNotDisposeNotifierModel implements Disposable, Identifiable<Integer> {
+final class Arez_DisposeNotifierDisabledCascadeDisposeModel extends DisposeNotifierDisabledCascadeDisposeModel implements Disposable, Identifiable<Integer> {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;
 
-  Arez_MultiPreDisposeNotDisposeNotifierModel() {
+  Arez_DisposeNotifierDisabledCascadeDisposeModel() {
     super();
     final ArezContext $$arezv$$_context = Arez.context();
     final int $$arezv$$_id = ++$$arezi$$_nextId;
-    final String $$arezv$$_name = Arez.areNamesEnabled() ? "MultiPreDisposeNotDisposeNotifierModel." + $$arezv$$_id : null;
-    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "MultiPreDisposeNotDisposeNotifierModel", $$arezv$$_id, $$arezv$$_name, this::$$arezi$$_nativeComponentPreDispose ) : null;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "DisposeNotifierDisabledCascadeDisposeModel." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "DisposeNotifierDisabledCascadeDisposeModel", $$arezv$$_id, $$arezv$$_name, this::$$arezi$$_nativeComponentPreDispose ) : null;
     this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNamesEnabled() ? $$arezv$$_name : null, $$arezv$$_id, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNativeComponentsEnabled() ? null : this::$$arezi$$_preDispose, null, null, false, false, false );
     this.$$arezi$$_kernel.componentConstructed();
     this.$$arezi$$_kernel.componentReady();
@@ -37,8 +37,7 @@ final class Arez_MultiPreDisposeNotDisposeNotifierModel extends MultiPreDisposeN
   }
 
   private void $$arezi$$_preDispose() {
-    super.doStuff();
-    super.foo();
+    Disposable.dispose( _myObject );
   }
 
   private void $$arezi$$_nativeComponentPreDispose() {