+++ processor/src/test/fixtures/input/com/example/cascade_dispose/DisposeNotifierDisabledCascadeDisposeModel.java
@@ -1,0 +1,13 @@
+package com.example.cascade_dispose;
+
+import arez.Disposable;
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+import arez.annotations.Feature;
+
+@ArezComponent( disposeNotifier = Feature.DISABLE )
+abstract class DisposeNotifierDisabledCascadeDisposeModel
+{
+  @CascadeDispose
+  final Disposable _myObject = null;
+}